+++ src/fhir/STU3/sdc-export.js
@@ -2,17 +2,11 @@
  * A package to handle FHIR Questionnaire and SDC (STU2) Questionnaire and QuestionnaireResponse for LForms
  *
  * FHIR Questionnaire:
- * https://www.hl7.org/fhir/questionnaire.html
- *
+ * https://hl7.org/fhir/STU3/questionnaire.html
+ * 
  * STU2 SDC Ballot:
- * http://hl7.org/fhir/us/sdc/sdc-questionnaire.html
- * http://hl7.org/fhir/us/sdc/sdc-questionnaireresponse.html
+ * https://hl7.org/fhir/us/sdc/STU2/
  *
- * It provides the following functions:
- * convertLFormsToQuestionnaire()
- * -- Convert existing LOINC panels/forms data in LForms format into FHIR (standard or SDC) Questionnaire data
- * convertLFormsToQuestionnaireResponse()
- * -- Generate FHIR (standard or SDC) QuestionnaireResponse data from captured data in LForms
  */
 var fhirVersionNum = '3.0';